fix + addition
some UI improvements, also made the base for improved user register

+++ pages/auth/redirect/[redirType].tsx
@@ -10,17 +10,32 @@ import { useEffect, useState } from "react";
 import { useRouter } from "next/router";
 
 const ParentContainer = styled.div`
-width: 230px;
+width: 250px;
 display:flex;
-margin-left:auto;
-margin-right:auto;
 color: ${props => props.theme.Body.TextColorLevels[2]};
 font-family: 'Open Sans', sans-serif;
-padding: 30px;
+padding: 30px 50px;
 flex-direction:column;
-margin-top: 50px;
+position: absolute;
+left: 50%;
+top: 300px;
+transform: translate(-50%,-50%);
 justify-content:center;
-height: calc(100vh - 100px);
+border-radius: 20px;
+align-items: flex-start;
+@media(max-width: 590px) {
+    width: calc(100% - 100px);
+    padding: 0px 50px;
+    height: 100vh;
+    margin-top: 0;
+    border: none;
+}
+& h1 {
+    font-size: 20px;
+    margin: 0px;
+    margin-top: 20px;
+    margin-bottom: 10px;
+}
 `
 
 const PolicyContainer = styled.div`
@@ -42,44 +57,49 @@ interface loginResp {
 
 export default function Login() {
     const isDark = useSelector<StateType, boolean>(state => state.theme);
-    const [isLoaded, setLoad] = useState(false)
+
     const [redirResult, setresult] = useState<loginResp>()
     const router = useRouter()
 
     useEffect(() => {
-        setLoad(true)
-    }, [])
-    useEffect(() => {
         if (router.query.hasOwnProperty("code")) {
             fetch(`/api/oauth2/${router.query.redirType}`, { method: "POST", body: JSON.stringify(router.query) }).then((data) => data.json()).then((jsn) => {
-                if (jsn.status == "Success") {
+                if (jsn.status == "Success" && jsn.accountExists) {
                     localStorage.setItem("tk", jsn.token)
+                    setTimeout(() => { router.replace("/") }, 1000)
                 }
                 setresult(jsn)
-                setTimeout(() => { router.replace("/") }, 1000)
             })
         }
-    }, [router.isReady])
+    }, [router.isReady, router])
 
     return (
         <>
             <ThemeProvider theme={isDark ? DarkTheme : LightTheme} >
-                <GlobalStyle />
-                {isLoaded ?
-                    <ParentContainer>
-                        {redirResult ? <>
-                            {redirResult.status == "Success" ? redirResult.accountExists ? "로그인되었습니다" : "가입이 완료되었습니다" : "오류가 발생했습니다"}
-                            <PolicyContainer>
-                                {redirResult.detail ? "Detail:" + redirResult.detail : "1초 후 자동으로 리다이렉트됩니다"}
-                            </PolicyContainer>
 
-                        </> :
+                <GlobalStyle />
+                <Header currentPage="login" />
+                {redirResult ?
+                    <ParentContainer>
+                        {redirResult.status == "Success" ? redirResult.accountExists ?
                             <>
-                                잠시만 기다려주세요..
-                            </>}
+                                로그인되었습니다
+                            </>
+                            : <>
+                                <h1>처음 뵙겠습니다</h1>
+
+                            </>
+                            :
+                            <>
+                                오류가 발생했습니다
+                            </>
+                        }
+                        <PolicyContainer>
+                            {redirResult.detail ? "Detail:" + redirResult.detail : ""}
+                        </PolicyContainer>
 
                     </ParentContainer>
-                    : <></>}
+                    : <>잠시만 기다려주세요..</>}
             </ThemeProvider>
         </>
     )
